move json schema tests to json module
