+++ gbnf/src/lib.rs
@@ -733,548 +733,6 @@ mod tests {
     use super::*;
 
     #[test]
-    fn simple_json_schema_boolean() {
-        let schema = r#"
-    {
-        "$id": "https://example.com/enumerated-values.schema.json",
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "title": "Enumerated Values",
-        "type": "boolean"
-    }
-            "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $id: https://example.com/enumerated-values.schema.json
-# $schema: https://json-schema.org/draft/2020-12/schema
-# title: Enumerated Values
-################################################
-
-root ::= boolean ws
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        );
-    }
-
-    #[test]
-    fn simple_json_schema_number() {
-        let schema = r#"
-    {
-        "$id": "https://example.com/enumerated-values.schema.json",
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "title": "Enumerated Values",
-        "type": "number"
-    }
-            "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $id: https://example.com/enumerated-values.schema.json
-# $schema: https://json-schema.org/draft/2020-12/schema
-# title: Enumerated Values
-################################################
-
-root ::= number ws
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        );
-    }
-
-    #[test]
-    fn simple_json_schema_string() {
-        let schema = r#"
-    {
-        "$id": "https://example.com/enumerated-values.schema.json",
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "title": "Enumerated Values",
-        "type": "string"
-    }
-            "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $id: https://example.com/enumerated-values.schema.json
-# $schema: https://json-schema.org/draft/2020-12/schema
-# title: Enumerated Values
-################################################
-
-root ::= string ws
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        );
-    }
-
-    #[test]
-    fn simple_json_schema_basic_object() {
-        let schema = r#"
-    {
-        "$id": "https://example.com/enumerated-values.schema.json",
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "title": "Enumerated Values",
-        "type": "object",
-        "properties": {
-            "a": {
-                "type": "boolean"
-            },
-            "b": {
-                "type": "number"
-            },
-            "c": {
-                "type": "string"
-            }
-        }
-    }
-            "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $id: https://example.com/enumerated-values.schema.json
-# $schema: https://json-schema.org/draft/2020-12/schema
-# title: Enumerated Values
-################################################
-
-symbol1-a-value ::= boolean ws
-symbol2-b-value ::= number ws
-symbol3-c-value ::= string ws
-root ::= "{" ws "\"a\"" ws ":" ws symbol1-a-value "," ws "\"b\"" ws ":" ws symbol2-b-value "," ws "\"c\"" ws ":" ws symbol3-c-value "}" ws
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        )
-    }
-
-    #[test]
-    fn simple_json_schema_nested_object() {
-        let schema = r#"
-    {
-        "$id": "https://example.com/enumerated-values.schema.json",
-        "$schema": "https://json-schema.org/draft/2020-12/schema",
-        "title": "Enumerated Values",
-        "type": "object",
-        "properties": {
-            "a": {
-                "type": "boolean"
-            },
-            "b": {
-                "type": "number"
-            },
-            "c": {
-                "type": "object",
-                "properties": {
-                    "x": {
-                        "type": "boolean"
-                    },
-                    "y": {
-                        "type": "number"
-                    },
-                    "z": {
-                        "type": "string"
-                    }
-                }
-            }
-        }
-    }
-            "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $id: https://example.com/enumerated-values.schema.json
-# $schema: https://json-schema.org/draft/2020-12/schema
-# title: Enumerated Values
-################################################
-
-symbol1-a-value ::= boolean ws
-symbol2-b-value ::= number ws
-symbol4-x-value ::= boolean ws
-symbol5-y-value ::= number ws
-symbol6-z-value ::= string ws
-symbol3-c-value ::= "{" ws "\"x\"" ws ":" ws symbol4-x-value "," ws "\"y\"" ws ":" ws symbol5-y-value "," ws "\"z\"" ws ":" ws symbol6-z-value "}" ws
-root ::= "{" ws "\"a\"" ws ":" ws symbol1-a-value "," ws "\"b\"" ws ":" ws symbol2-b-value "," ws "\"c\"" ws ":" ws symbol3-c-value "}" ws
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        )
-    }
-
-    #[test]
-    fn simple_json_schema_oneof() {
-        let schema = r#"
-   {
-	"$schema": "https://json-schema.org/draft/2019-09/schema",
-	"oneOf": [
-      {
-        "type" : "object",
-        "properties" : {
-            "firstName" : {
-                "type" : "string"
-            },
-            "lastName" : {
-                "type" : "string"
-            },
-            "sport" : {
-                "type" : "string"
-            }
-          }
-      },
-      {
-        "type" : "number"
-      }
-    ]
-}
-            "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $schema: https://json-schema.org/draft/2019-09/schema
-################################################
-
-symbol2-firstName-value ::= string ws
-symbol3-lastName-value ::= string ws
-symbol4-sport-value ::= string ws
-symbol-1-oneof-0 ::= "{" ws "\"firstName\"" ws ":" ws symbol2-firstName-value "," ws "\"lastName\"" ws ":" ws symbol3-lastName-value "," ws "\"sport\"" ws ":" ws symbol4-sport-value "}" ws
-symbol-5-oneof-1 ::= number ws
-root ::= symbol-1-oneof-0 | symbol-5-oneof-1
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#,
-        )
-    }
-
-    #[test]
-    fn simple_json_schema_enum() {
-        let schema = r#"
-        {
-         "$schema": "https://json-schema.org/draft/2019-09/schema",
-         "enum": [
-              "red",
-              "amber",
-              "green"
-         ]
-     }
-                 "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $schema: https://json-schema.org/draft/2019-09/schema
-################################################
-
-root ::= "\"red\"" | "\"amber\"" | "\"green\""
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        )
-    }
-
-    #[test]
-    fn simple_json_schema_value_string() {
-        let schema = r#"
-        {
-         "$schema": "https://json-schema.org/draft/2019-09/schema",
-         "const": "red"
-     }
-                 "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $schema: https://json-schema.org/draft/2019-09/schema
-################################################
-
-root ::= "\"red\""
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        )
-    }
-
-    #[test]
-    fn simple_json_schema_value_number() {
-        let schema = r#"
-        {
-         "$schema": "https://json-schema.org/draft/2019-09/schema",
-         "const": 42
-     }
-                 "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $schema: https://json-schema.org/draft/2019-09/schema
-################################################
-
-root ::= "42"
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        )
-    }
-
-    #[test]
-    fn simple_json_schema_value_boolean() {
-        let schema = r#"
-        {
-         "$schema": "https://json-schema.org/draft/2019-09/schema",
-         "const": true
-     }
-                 "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $schema: https://json-schema.org/draft/2019-09/schema
-################################################
-
-root ::= "true"
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        )
-    }
-
-    #[test]
-    fn simple_json_schema_array() {
-        let schema = r#"
-        {
-         "$schema": "https://json-schema.org/draft/2019-09/schema",
-         "type": "array",
-         "items": {
-             "type": "string"
-         }
-     }
-                 "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $schema: https://json-schema.org/draft/2019-09/schema
-################################################
-
-symbol1-item ::= string ws
-root ::= "[" ws symbol1-item* ws "]" ws
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        )
-    }
-
-    #[test]
-    fn simple_json_kitchen_sink() {
-        let schema = r#"
-        {
-            "$schema": "https://json-schema.org/draft/2019-09/schema",
-            "type": "object",
-            "properties": {
-                "name": {
-                    "type": "string"
-                },
-                "age": {
-                    "type": "number"
-                },
-                "usesAI": {
-                    "type": "boolean"
-                },
-                "favoriteAnimal": {
-                    "enum": [
-                        "dog",
-                        "cat",
-                        "none"
-                    ]
-                },
-                "currentAIModel": {
-                    "oneOf": [
-                        {
-                            "type": "object",
-                            "properties": {
-                                "type": {
-                                    "const": "hugging_face"
-                                },
-                                "name": {
-                                    "type": "string"
-                                }
-                            }
-                        },
-                        {
-                            "type": "object",
-                            "properties": {
-                                "type": {
-                                    "const": "openai"
-                                }
-                            }
-                        }
-                    ]
-                },
-                "favoriteColors": {
-                    "type": "array",
-                    "items": {
-                        "type": "string"
-                    }
-                }
-            }
-        }
-                 "#;
-        let g = Grammar::from_json_schema(schema).unwrap();
-        let s = g.to_string();
-
-        pretty_assertions::assert_eq!(
-            s,
-            r#"################################################
-# DYNAMICALLY GENERATED JSON-SCHEMA GRAMMAR
-# $schema: https://json-schema.org/draft/2019-09/schema
-################################################
-
-symbol1-name-value ::= string ws
-symbol2-age-value ::= number ws
-symbol3-usesAI-value ::= boolean ws
-symbol4-favoriteAnimal-value ::= "\"dog\"" | "\"cat\"" | "\"none\""
-symbol7-type-value ::= "\"hugging_face\""
-symbol8-name-value ::= string ws
-symbol-6-oneof-0 ::= "{" ws "\"type\"" ws ":" ws symbol7-type-value "," ws "\"name\"" ws ":" ws symbol8-name-value "}" ws
-symbol10-type-value ::= "\"openai\""
-symbol-9-oneof-1 ::= "{" ws "\"type\"" ws ":" ws symbol10-type-value "}" ws
-symbol5-currentAIModel-value ::= symbol-6-oneof-0 | symbol-9-oneof-1
-symbol12-item ::= string ws
-symbol11-favoriteColors-value ::= "[" ws symbol12-item* ws "]" ws
-root ::= "{" ws "\"name\"" ws ":" ws symbol1-name-value "," ws "\"age\"" ws ":" ws symbol2-age-value "," ws "\"usesAI\"" ws ":" ws symbol3-usesAI-value "," ws "\"favoriteAnimal\"" ws ":" ws symbol4-favoriteAnimal-value "," ws "\"currentAIModel\"" ws ":" ws symbol5-currentAIModel-value "," ws "\"favoriteColors\"" ws ":" ws symbol11-favoriteColors-value "}" ws
-
-###############################
-# Primitive value type symbols
-###############################
-boolean ::= "true" | "false" ws
-integer ::= ("-"? ([0-9] | [1-9] [0-9]*)) ws
-null ::= "null" ws
-number ::= ("-"? ([0-9] | [1-9] [0-9]*)) ("." [0-9]+)? ([eE] [-+]? [0-9]+)? ws
-string ::= "\"" ([^"\\] | "\\" (["\\/bfnrt] | "u" [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F] [0-9a-fA-F]))* "\"" ws
-ws ::= [ \t\n]*
-"#
-        )
-    }
-
-    #[test]
     fn simple_0() {
         // root ::= "yes" | "no"
         let g = Grammar {
